+++ src/Filesystem.php
@@ -37,6 +37,19 @@ class Filesystem extends BaseFilesystem
     }
 
     /**
+     * Concatenates various paths together, adding the right slash term
+     * @param string $paths Various paths to be concatenated
+     * @return string The path concatenated
+     * @since 1.4.5
+     */
+    public function concatenate(...$paths)
+    {
+        $end = array_pop($paths);
+
+        return implode('', array_map([$this, 'addSlashTerm'], $paths)) . $end;
+    }
+
+    /**
      * Creates a file. Alias for `mkdir()` and `file_put_contents()`.
      *
      * It also recursively creates the directory where the file will be created.
@@ -199,7 +212,7 @@ class Filesystem extends BaseFilesystem
             return $endPath;
         }
 
-        return $this->addSlashTerm($startPath) . $endPath;
+        return $this->concatenate($startPath, $endPath);
     }
 
     /**